api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic>=2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
# Minimal RAG setup - only essential packages
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
# Remove heavy dependencies for now
# tavily-python==0.3.0
# chromadb==0.4.18
# sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1